+++ pkg/linter/test/rules/avoid_field_initializers_in_non_const_classes_test.dart
@@ -63,6 +63,7 @@ augment class A {
     );
   }
 
+  @SkippedTest() // TODO(scheglov): implement augmentation
   test_augmentedClass_augmentedField() async {
     newFile('$testPackageLibPath/a.dart', r'''
 part 'test.dart';