+++ pkg/linter/test/rules/overridden_fields_test.dart
@@ -40,6 +40,7 @@ augment class A {
     await assertDiagnosticsInFile(b.path, [lint(45, 1)]);
   }
 
+  @SkippedTest() // TODO(scheglov): implement augmentation
   test_augmentedField() async {
     var a = newFile('$testPackageLibPath/a.dart', r'''
 part 'b.dart';