+++ runtime/bin/dart_embedder_api_impl.cc
@@ -6,7 +6,11 @@
 
 #include "bin/dartutils.h"
 #include "bin/eventhandler.h"
+#if defined(DART_IO_SECURE_SOCKET_DISABLED)
+#include "bin/io_service_no_ssl.h"
+#else  // defined(DART_IO_SECURE_SOCKET_DISABLED)
 #include "bin/io_service.h"
+#endif  // defined(DART_IO_SECURE_SOCKET_DISABLED)
 #include "bin/isolate_data.h"
 #include "bin/process.h"
 #include "bin/secure_socket_filter.h"