+++ pkg/analyzer/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Remove deprecated `RecordType.sortedNamedTypes`.
 * Remove `ElementLocation` class; its values are not returned anymore.
 * Remove deprecated `AnalysisContext.analysisOptions`.
+* Remove deprecated `PromotableElement` and `PromotableFragment`.
 * `LibraryFragment` is not `Annotatable` anymore.
 * Stop implementing `ConstantEvaluationTarget` by `ElementAnnotation`.
   This is an internal interface that should not have been exposed.